+++ CHANGELOG.md
@@ -48,6 +48,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `RxReusableTableViewCell` class for working reactively with a table view.
 - Reactive extension with `UIViewController` lifecycle events.
 - Implement stubs for `LaunchListResponse'`, `LaunchListItem`, `LaunchResponse`, `LaunchServiceProviderResponse`, `PadResponse`, `LocationResponse`.
+- `LaunchListDataSource` and `LaunchListTableViewCell`.
+- `default_rocket_launch_image.jpg` to the Assets.
 
 ### Changed